requirements.txt
git+https://github.com/huggingface/transformers
gradio
torch
pillow
accelerate>=0.26.0